requirements.txt
openai
whisper
moviepy
ffmpeg-python
python-dotenv

